Reorg example into files
